# lib/api-zod/src/generated/types/fileItem.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Smart Private Cloud Storage Agent API
 * OpenAPI spec version: 0.1.0
 */

export interface FileItem {
  id: string;
  name: string;
  path: string;
  diskId: string;
  sizeBytes: number;
  mimeType?: string;
  category: string;
  aiSuggestedPath?: string;
  aiApproved: boolean;
  thumbnail?: string;
  createdAt: string;
  modifiedAt: string;
  accessedAt?: string;
}