# lib/api-zod/src/generated/types/oAuthConnection.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Smart Private Cloud Storage Agent API
 * OpenAPI spec version: 0.1.0
 */
import type { OAuthConnectionProvider } from "./oAuthConnectionProvider";

export interface OAuthConnection {
  id: string;
  provider: OAuthConnectionProvider;
  connected: boolean;
  username?: string;
  avatar?: string;
  connectedAt?: string;
}