# lib/api-zod/src/generated/types/suggestStorageBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Smart Private Cloud Storage Agent API
 * OpenAPI spec version: 0.1.0
 */

export interface SuggestStorageBody {
  fileIds: string[];
  description?: string;
}